feat: enhance workout session management with user authorization and validation

+++ src/TrainingService/Workouts/DeleteWorkoutById/DeleteWorkoutByIdCommandHandler.cs
@@ -24,6 +24,9 @@ public class DeleteWorkoutByIdCommandHandler(IWorkoutRepository repository)
         if (workout is null)
             throw new NotFoundException($"Workout with id '{command.WorkoutId}' not found.");
         
+        if(workout.CreatorId != command.UserId)
+            throw new ForbiddenException($"User '{command.UserId}' is not authorized to delete this workout.");
+        
         await repository.DeleteAsync(workout, cancellationToken);
         
         return true;
